Module storage_proofs_update::constants

Structs

POSEIDON_CONSTANTS_GEN_RANDOMNESS

Constants

ALLOWED_SECTOR_SIZES

HASH_TYPE_GEN_RANDOMNESS

SECTOR_SIZE_1_KIB

SECTOR_SIZE_2_KIB

SECTOR_SIZE_4_KIB

SECTOR_SIZE_8_KIB

SECTOR_SIZE_8_MIB

SECTOR_SIZE_16_KIB

SECTOR_SIZE_16_MIB

SECTOR_SIZE_32_GIB

SECTOR_SIZE_32_KIB

SECTOR_SIZE_64_GIB

SECTOR_SIZE_512_MIB

Functions

apex_leaf_count

challenge_count

challenge_count_poseidon

h_default

Returns the h for the given sector-size. The h value is the number of high bits taken from each challenge c. For production use, it was determined to use the value at index 3, which translates to a value of 10 for production sector sizes.

h_select

Returns the bit pattern to select the h value from the list of hard-coded h values within the circuit.

hs

Returns the h values allowed for the given sector-size. Each h value is a possible number of high bits taken from each challenge c. A single value of h = hs[i] is taken from hs for each proof; the circuit takes h_select = 2^i as a public input.

partition_count

validate_tree_r_shape

Type Aliases

TreeD

TreeDArity

TreeDDomain

TreeDHasher

TreeDStore

TreeRBaseTree

TreeRDomain

TreeRHasher